--- src/.eslintcache
@@ -1,1 +1,0 @@
-[{"F:\\YT\\Projects\\react-responsive-navbar\\src\\index.js":"1","F:\\YT\\Projects\\react-responsive-navbar\\src\\App.js":"2","F:\\YT\\Projects\\react-responsive-navbar\\src\\reportWebVitals.js":"3","F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\NavBar.js":"4","F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\Home.js":"5","F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\About.js":"6","F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\Blog.js":"7","F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\Contact.js":"8"},{"size":500,"mtime":499162500000,"results":"9","hashOfConfig":"10"},{"size":764,"mtime":1607410453403,"results":"11","hashOfConfig":"10"},{"size":362,"mtime":499162500000,"results":"12","hashOfConfig":"10"},{"size":2056,"mtime":1607411838321,"results":"13","hashOfConfig":"10"},{"size":125,"mtime":1607410296057,"results":"14","hashOfConfig":"10"},{"size":127,"mtime":1607410314143,"results":"15","hashOfConfig":"10"},{"size":125,"mtime":1607410328870,"results":"16","hashOfConfig":"10"},{"size":134,"mtime":1607410345762,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"nq4fi7",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"F:\\YT\\Projects\\react-responsive-navbar\\src\\index.js",[],["35","36"],"F:\\YT\\Projects\\react-responsive-navbar\\src\\App.js",[],"F:\\YT\\Projects\\react-responsive-navbar\\src\\reportWebVitals.js",[],"F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\NavBar.js",[],"F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\Home.js",[],"F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\About.js",[],"F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\Blog.js",[],"F:\\YT\\Projects\\react-responsive-navbar\\src\\components\\Pages\\Contact.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
